fix(explore): constrain item height to prevent layout overflow in CoffeeShopListScreen
